kg_queries/scripts/requirements.txt
neo4j==5.14.0
openai==1.12.0
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
